Group command type definitions and helpers
