--- lib/type-helpers/command-builder.ts
@@ -1,5 +1,0 @@
-import { CommandBuilder, FunctionCommandBuilder } from '../types'
-
-export function isFunctionCommandBuilder (builder: CommandBuilder): builder is FunctionCommandBuilder {
-  return typeof builder === 'function'
-}
