+++ lib/type-helpers/index.ts
@@ -1,2 +1,1 @@
 export * from './completion-function'
-export * from './command-builder'
