--- lib/types/command-builder.ts
@@ -1,8 +1,0 @@
-import { YargsInstance } from './yargs-instance'
-
-// To be completed later with other CommandBuilder flavours
-export type CommandBuilder = FunctionCommandBuilder
-
-export interface FunctionCommandBuilder {
-  (y: YargsInstance): YargsInstance
-}
