--- lib/types/command-handler.ts
@@ -1,5 +1,0 @@
-import { CommandBuilder } from './command-builder'
-
-export interface CommandHandler {
-  builder: CommandBuilder
-}
