--- lib/types/command-instance.ts
@@ -1,8 +1,0 @@
-import { CommandHandler } from './command-handler'
-import { Dictionary } from './dictionary'
-
-/** Instance of the command module. */
-export interface CommandInstance {
-  getCommandHandlers (): Dictionary<CommandHandler>
-  getCommands (): string[]
-}
